requirements.txt
aws-cdk.core
aws_cdk.aws_guardduty
-e .
